+++ pkg/hub/api/clusters/clusters.go
@@ -1,7 +1,6 @@
 package clusters
 
 import (
-	"fmt"
 	"net/http"
 
 	"github.com/kobsio/kobs/pkg/hub/store"
@@ -51,20 +50,13 @@ func (router *Router) getNamespaces(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	var groupedNamespaces map[string][]shared.Namespace
-	groupedNamespaces = make(map[string][]shared.Namespace)
+	var uniqueNamespaces []string
 
 	for _, namespace := range namespaces {
-		key := fmt.Sprintf("%s (%s)", namespace.Cluster, namespace.Satellite)
-
-		if _, ok := groupedNamespaces[key]; ok {
-			groupedNamespaces[key] = append(groupedNamespaces[key], namespace)
-		} else {
-			groupedNamespaces[key] = []shared.Namespace{namespace}
-		}
+		uniqueNamespaces = appendIfMissing(uniqueNamespaces, namespace.Namespace)
 	}
 
-	render.JSON(w, r, groupedNamespaces)
+	render.JSON(w, r, uniqueNamespaces)
 }
 
 func (router *Router) getResources(w http.ResponseWriter, r *http.Request) {