+++ plugins/app/src/components/applications/utils/interfaces.ts
@@ -2,7 +2,7 @@ export interface IOptions {
   all: boolean;
   clusterIDs: string[];
   external: string;
-  namespaceIDs: string[];
+  namespaces: string[];
   page: number;
   perPage: number;
   searchTerm: string;