+++ plugins/app/src/resources/clusters.ts
@@ -9,19 +9,6 @@ export interface ICluster {
   updatedAt: number;
 }
 
-export interface INamespaces {
-  [cluster: string]: INamespace[];
-}
-
-export interface INamespace {
-  id: string;
-  namespace: string;
-  cluster: string;
-  satellite: string;
-  clusterID: string;
-  updatedAt: number;
-}
-
 export interface IResource {
   id: string;
   description: string;